Add files via upload
The Employee Management System is a database application that allows businesses to manage employee data efficiently. It provides functionalities such as storing employee information, managing job titles, tracking salaries, departments, and employee performance, and generating various reports. This type of system can be critical for HR and management teams to keep track of an organization's workforce.